Support arguments for messages
Support arguments for messages, similar to the Grails message tag.
Arguments can be given to $L either as additional argument values, or as
an array of values.

Messages containing positional parameters are converted to arrays before
they are added to the JavaScript map to keep the $L function simple.

Added test, updated README